requirements.txt
pdfplumber
faiss-cpu
requests
numpy
streamlit